backend/app/pyproject.toml
[tool.poetry]
name = "vktest"
version = "0.1.0"
description = ""
authors = ["DarkLorianPrime <lolpokens@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
